+++ internal/bot/components/bios/systemCommands.go
@@ -50,7 +50,7 @@ func (b *Bios) SystemImportMember(ctx *route.MessageContext) error {
 
 	for _, x := range accBios {
 		if strings.EqualFold(x.SysMemberID, memberId) {
-			_, err := ctx.SendMessageString(ctx.Message.ChannelID, "This member ID has already been imported!")
+			err = ctx.SendErrorMessage("This member ID has already been imported!")
 			return err
 		}
 	}
@@ -59,8 +59,7 @@ func (b *Bios) SystemImportMember(ctx *route.MessageContext) error {
 	systemInfo, err := pluralkit.SystemByDiscordAccount(ctx.Message.Author.ID)
 	if err != nil {
 		if errors.Is(err, pluralkit.ErrorAccountHasNoSystem) {
-			_, err := ctx.SendMessageString(ctx.Message.ChannelID, "Your Discord account has no PluralKit systems associated "+
-				"with it.")
+			err = ctx.SendErrorMessage("Your Discord account has no PluralKit systems associated with it.")
 			return err
 		}
 		return err
@@ -70,8 +69,8 @@ func (b *Bios) SystemImportMember(ctx *route.MessageContext) error {
 	systemMembers, err := pluralkit.MembersBySystemId(systemInfo.Id)
 	if err != nil {
 		if errors.Is(err, pluralkit.ErrorMemberListPrivate) {
-			_, err := ctx.SendMessageString(ctx.Message.ChannelID, "Your system has the member list set to "+
-				"**private**. Please set this to public and try again (HTTP 403)")
+			err = ctx.SendErrorMessage("Your system has the member list set to **private**. Please set this " +
+				"to public and try again (you can set it to private again afterwards)")
 			return err
 		}
 		return err
@@ -86,9 +85,8 @@ func (b *Bios) SystemImportMember(ctx *route.MessageContext) error {
 	}
 
 	if pkMember == nil {
-		_, err := ctx.SendMessageString(ctx.Message.ChannelID, "Your system has has no member with the given "+
-			"ID. If you're sure there's a registered member with this ID, make sure the member visibility privacy "+
-			"level is set to "+"**public**.")
+		err = ctx.SendErrorMessage("Your system has has no member with the given ID. If you're sure there's " +
+			"a registered member with this ID, make sure the member visibility privacy level is set to **public**.")
 		return err
 	}