+++ internal/bot/components/chatchart/commands.go
@@ -11,25 +11,21 @@ func (c *ChatChart) Trigger(ctx *route.MessageContext) error {
 	channelId := ctx.Arguments["channel"].(string)
 
 	if _, err := ctx.Session.Channel(channelId); err != nil {
-		_, err := ctx.SendMessageString(ctx.Message.ChannelID, "Invalid channel")
+		err = ctx.SendErrorMessage("Invalid channel")
 		if err != nil {
 			return err
 		}
 	}
 
-	var msg string
-
 	if tools.IsStringInSlice(channelId, config.ChatChartChannelExclusions) {
-		msg = "This channel has been excluded from chat chart indexing."
+		return ctx.SendErrorMessage("This channel has been excluded from chat chart indexing.")
 	} else {
 		c.queue <- collectionIntent{
 			ChannelId: channelId,
 			Ctx:   ctx,
 		}
-		msg = "Task queued. You'll be pinged when collection is complete and a chart is ready."
+		_, err := ctx.SendMessageString(ctx.Message.ChannelID, "Task queued. You'll be pinged when collection" +
+			" is complete and a chart is ready.")
+		return err
 	}
-
-	_, err := ctx.SendMessageString(ctx.Message.ChannelID, msg)
-	return err
-
 }